# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/stock/balances": {
      "get": {
        "operationId": "StockController_balances",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "belowMin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Saldos por filial/local (disponível = físico - reservado)",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/movements": {
      "get": {
        "operationId": "StockController_movements",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Movimentações imutáveis",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/alerts": {
      "get": {
        "operationId": "StockController_alerts",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Itens abaixo do estoque mínimo",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/receive": {
      "post": {
        "operationId": "StockController_receive",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "locationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "supplierId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "reference": {
                    "type": "string",
                    "maxLength": 120,
                    "nullable": true
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "quantity": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "maximum": 1000000
                        },
                        "unitCostCents": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1000000000
                        }
                      },
                      "required": [
                        "productId",
                        "quantity",
                        "unitCostCents"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 200
                  }
                },
                "required": [
                  "branchId",
                  "items"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Entrada de mercadoria (custo médio)",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/adjust": {
      "post": {
        "operationId": "StockController_adjust",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "locationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "productId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "quantityDelta": {
                    "type": "integer"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "ADJUSTMENT",
                      "LOSS",
                      "RETURN"
                    ],
                    "default": "ADJUSTMENT"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 300
                  },
                  "overrideNegative": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "branchId",
                  "productId",
                  "quantityDelta",
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajuste/perda/devolução (negativo exige override auditado)",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/inventory": {
      "post": {
        "operationId": "StockController_inventory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 300
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "countedQty": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "productId",
                        "countedQty"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 500
                  }
                },
                "required": [
                  "branchId",
                  "reason",
                  "items"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inventário com reconciliação",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/reserve": {
      "post": {
        "operationId": "StockController_reserve",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "productId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "quantity": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "maximum": 10000
                  },
                  "locationId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "orderId",
                  "productId",
                  "quantity"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reserva peça para OS",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/consume": {
      "post": {
        "operationId": "StockController_consume",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reservationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "quantity": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "maximum": 10000
                  }
                },
                "required": [
                  "reservationId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Baixa (consumo efetivo) de peça reservada",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/release": {
      "post": {
        "operationId": "StockController_release",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reservationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 300
                  }
                },
                "required": [
                  "reservationId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Libera reserva",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/transfers": {
      "get": {
        "operationId": "StockController_transfers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "IN_TRANSIT",
                "RECEIVED",
                "RECEIVED_WITH_DIFF",
                "CANCELED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transferências entre filiais",
        "tags": [
          "stock"
        ]
      },
      "post": {
        "operationId": "StockController_createTransfer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sourceBranchId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "destBranchId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 500,
                    "nullable": true
                  },
                  "lines": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "quantity": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "maximum": 1000000
                        }
                      },
                      "required": [
                        "productId",
                        "quantity"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 200
                  }
                },
                "required": [
                  "sourceBranchId",
                  "destBranchId",
                  "lines"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Envia transferência (sai da origem, fica em trânsito)",
        "tags": [
          "stock"
        ]
      }
    },
    "/api/v1/stock/transfers/{id}/receive": {
      "post": {
        "operationId": "StockController_receiveTransfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notes": {
                    "type": "string",
                    "maxLength": 500,
                    "nullable": true
                  },
                  "lines": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "lineId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "quantityReceived": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1000000
                        }
                      },
                      "required": [
                        "lineId",
                        "quantityReceived"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 200
                  }
                },
                "required": [
                  "lines"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirma recebimento com diferenças",
        "tags": [
          "stock"
        ]
      }
    },
...
  },
  "info": {
    "title": "OrdemCerta API",
    "description": "API REST v1 — gestão de assistências técnicas. Erros: {code,message,details,requestId}.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}